fix: audit — wire governance, fix metric bugs, register news API, improve knowledge graph
- Fix experiment_proposals._get_current_metrics: use 'last' not 'latest' key from trend()
- Fix meta_metrics.dashboard: cast float to int for days parameter
- Register news_feed_api blueprint on hub Flask app (/api/news/*)
- Wire governance validation into preToolUse hook (check-tool-safety.ps1)
- Add governance-audit scheduler callback (11th builtin task, weekly)
- Fix knowledge_graph tag parsing for JSON-encoded tag arrays
- Add domain stop words to reduce noisy gap detection
- Update test assertions for 11 builtin tasks

+++ engine/nexus/meta_metrics.py
@@ -587,7 +587,7 @@ class MetaMetrics:
 
             for name in metric_names:
                 current = self._latest_value(name)
-                short_trend = self.trend(name, days=max(hours / 24, 1))
+                short_trend = self.trend(name, days=int(max(hours / 24, 1)))
                 long_trend = self.trend(name, days=7)
                 short_arrow = self._trend_arrow(short_trend)
                 long_arrow = self._trend_arrow(long_trend)
